pyproject.toml
[tool.poetry]
name = "httprunner"
version = "3.1.6"
description = "Based on Httprunner transformation framework, the database operation is extended."
license = "Apache-2.0"
readme = "README.md"
authors = ["debugtalk <debugtalk@gmail.com>", "zhaochencheng <907779487@qq.com>"]

homepage = "https://github.com/zhaochencheng/httprunner_pro"
repository = "https://github.com/zhaochencheng/httprunner_pro"
documentation = "https://docs.httprunner.org"

keywords = ["HTTP", "api", "test", "requests", "locustio"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8"
]

include = ["docs/CHANGELOG.md"]

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.22.0"
pyyaml = "^5.1.2"
jinja2 = "^2.10.3"
pydantic = "^1.4"
loguru = "^0.4.1"
jmespath = "^0.9.5"
black = "^19.10b0"
pytest = "^5.4.2"
pytest-html = "^2.1.1"
sentry-sdk = "^0.14.4"
allure-pytest = {version = "^2.8.16", optional = true}
requests-toolbelt = {version = "^0.9.1", optional = true}
filetype = {version = "^1.0.7", optional = true}
locust = {version = "^1.0.3", optional = true}
Brotli = "^1.0.9"
pymongo = "^3.11.0"
PyMySQL = "^1.0.2"
redis-py-cluster = "^2.1.3"

[tool.poetry.extras]
allure = ["allure-pytest"]                  # pip install "httprunner[allure]", poetry install -E allure
upload = ["requests-toolbelt", "filetype"]  # pip install "httprunner[upload]", poetry install -E upload
locust = ["locust"]                         # pip install "httprunner[locust]", poetry install -E locust

[tool.poetry.dev-dependencies]
coverage = "^4.5.4"
uvicorn = "^0.11.3"
fastapi = "^0.49.0"

[tool.poetry.scripts]
httprunner = "httprunner.cli:main"
hrun = "httprunner.cli:main_hrun_alias"
hmake = "httprunner.cli:main_make_alias"
har2case = "httprunner.cli:main_har2case_alias"
locusts = "httprunner.ext.locust:main_locusts"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
